+++ src/consensus/params.h
@@ -108,7 +108,7 @@ struct Params {
     int MinBIP9WarningHeight;
     std::array<BIP9Deployment,MAX_VERSION_BITS_DEPLOYMENTS> vDeployments;
     /** Proof of work parameters */
-    uint256 powLimit;
+    uint16_t powLimit;
     bool fPowAllowMinDifficultyBlocks;
     /**
       * Enforce BIP94 timewarp attack mitigation. On testnet4 this also enforces
